Splits decorated_types.rs into separate files
