+++ full-service/src/api/account.rs
@@ -1,0 +1,17 @@
+use serde_derive::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Default, Debug, Clone)]
+pub struct JsonAccount {
+    pub object: String,
+    pub account_id: String,
+    pub name: String,
+    pub network_height: String,
+    pub local_height: String,
+    pub account_height: String,
+    pub is_synced: bool,
+    pub available_pmob: String,
+    pub pending_pmob: String,
+    pub main_address: String,
+    pub next_subaddress_index: String,
+    pub recovery_mode: bool,
+}
