+++ full-service/src/api/block.rs
@@ -1,0 +1,55 @@
+use mc_mobilecoind_json::data_types::{JsonTxOut, JsonTxOutMembershipElement};
+use serde_derive::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Default, Debug)]
+pub struct JsonBlock {
+    pub id: String,
+    pub version: String,
+    pub parent_id: String,
+    pub index: String,
+    pub cumulative_txo_count: String,
+    pub root_element: JsonTxOutMembershipElement,
+    pub contents_hash: String,
+}
+
+impl JsonBlock {
+    pub fn new(block: &mc_transaction_core::Block) -> Self {
+        let membership_element_proto =
+            mc_api::external::TxOutMembershipElement::from(&block.root_element);
+        Self {
+            id: hex::encode(block.id.clone()),
+            version: block.version.to_string(),
+            parent_id: hex::encode(block.parent_id.clone()),
+            index: block.index.to_string(),
+            cumulative_txo_count: block.cumulative_txo_count.to_string(),
+            root_element: JsonTxOutMembershipElement::from(&membership_element_proto),
+            contents_hash: hex::encode(block.contents_hash.0),
+        }
+    }
+}
+
+#[derive(Deserialize, Serialize, Default, Debug)]
+pub struct JsonBlockContents {
+    pub key_images: Vec<String>,
+    pub outputs: Vec<JsonTxOut>,
+}
+
+impl JsonBlockContents {
+    pub fn new(block_contents: &mc_transaction_core::BlockContents) -> Self {
+        Self {
+            key_images: block_contents
+                .key_images
+                .iter()
+                .map(|k| hex::encode(mc_util_serial::encode(k)))
+                .collect::<Vec<String>>(),
+            outputs: block_contents
+                .outputs
+                .iter()
+                .map(|txo| {
+                    let proto_txo = mc_api::external::TxOut::from(txo);
+                    JsonTxOut::from(&proto_txo)
+                })
+                .collect::<Vec<JsonTxOut>>(),
+        }
+    }
+}
