+++ full-service/src/api/mod.rs
@@ -2,9 +2,26 @@
 
 //! MobileCoin FullService Wallet API Layer.
 
-pub mod decorated_types;
+mod account;
+mod address;
+mod block;
 mod json_rpc_request;
 mod json_rpc_response;
+mod membership_proof;
+mod transaction_log;
+mod txo;
 mod wallet_api;
+mod wallet_status;
 
+pub use account::JsonAccount;
+pub use address::JsonAddress;
+pub use block::{JsonBlock, JsonBlockContents};
+pub use json_rpc_response::JsonBalanceResponse;
+pub use json_rpc_response::JsonCreateAccountResponse;
+pub use json_rpc_response::JsonSubmitResponse;
+pub use json_rpc_response::Response;
+pub use membership_proof::JsonProof;
+pub use transaction_log::JsonTransactionLog;
+pub use txo::JsonTxo;
 pub use wallet_api::{rocket, WalletApiState};
+pub use wallet_status::JsonWalletStatus;
