+++ full-service/src/db/account.rs
@@ -17,7 +17,7 @@ use mc_account_keys::{AccountKey, DEFAULT_SUBADDRESS_INDEX};
 use mc_crypto_digestible::{Digestible, MerlinTranscript};
 use mc_transaction_core::ring_signature::KeyImage;
 
-use crate::api::decorated_types::JsonAccount;
+use crate::api::JsonAccount; // The DB layer probably shouldn't know about JSON in the API layer...
 use crate::db::WalletDbError;
 use diesel::{
     prelude::*,
